Some new components and css
I made:

Some new components and routing to them: News, Selling Page, Cart and Contacts.

On Selling Page I add html and Css for the two options - with user, or when there is no user.

I made also a News Page. There the info for now is static.
I change the name of the site to - E-book. This you can see on the browser.
I made also a img_Books - there I save some pictures.